core/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==5.1.1
django-redis==5.4.0
djangorestframework==3.15.2
modeltranslation==0.25
redis==5.1.0
sqlparse==0.5.1
tzdata==2024.2
